Refactored PR
