+++ agr_indexer/pom.xml
@@ -27,10 +27,6 @@
 				</exclusion>
 			</exclusions>
 		</dependency>
-		<dependency>
-    		<groupId>io.quarkus</groupId>
-    		<artifactId>quarkus-rest-client</artifactId>
-  		</dependency>
 	</dependencies>
