+++ agr_indexer/src/main/java/org/alliancegenome/indexer/config/IndexerConfig.java
@@ -28,7 +28,7 @@ public enum IndexerConfig {
 	GoIndexer("go", GoIndexer.class, 4, 914, 914, 8, 1),
 	ModelIndexer("model", ModelIndexer.class, 4, 1500, 1426, 4, 1),
 
-	LiteratureIndexer("literature", LiteratureIndexer.class, 4, 5000, 5000, 4, 1),
+	LiteratureIndexer("literature", LiteratureIndexer.class, 4, 5000, 5000, 1, 1),
 	
 	// Curation Indexers
 	DiseaseAnnotationIndexer("diseaseAnnotation", DiseaseAnnotationCurationIndexer.class, 1, 1500, 1500, 2, 1),
